rename stripe directory as "renewals"
